QTi/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
redis==5.0.1
rclone==0.4.4
pytest==7.4.3
httpx==0.25.2
websockets==12.0
aiohttp==3.9.1
pandas==2.1.3
numpy==1.26.2
ccxt==4.1.87
ta-lib==0.4.28
python-telegram-bot==20.7 